Separate BlockMeta interfaces
