+++ core/server/worker/src/test/java/alluxio/worker/block/BlockMetadataManagerTest.java
@@ -21,6 +21,8 @@ import alluxio.exception.BlockDoesNotExistException;
 import alluxio.exception.ExceptionMessage;
 import alluxio.exception.WorkerOutOfSpaceException;
 import alluxio.worker.block.meta.BlockMeta;
+import alluxio.worker.block.meta.DefaultBlockMeta;
+import alluxio.worker.block.meta.DefaultTempBlockMeta;
 import alluxio.worker.block.meta.StorageDir;
 import alluxio.worker.block.meta.StorageTier;
 import alluxio.worker.block.meta.TempBlockMeta;
@@ -174,7 +176,7 @@ public final class BlockMetadataManagerTest {
   public void blockMeta() throws Exception {
     StorageDir dir = mMetaManager.getTier("HDD").getDir(0);
     TempBlockMeta tempBlockMeta =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
 
     // Empty storage
     assertFalse(mMetaManager.hasTempBlockMeta(TEST_TEMP_BLOCK_ID));
@@ -237,9 +239,9 @@ public final class BlockMetadataManagerTest {
     // create and add two temp block metas with same tier and dir to the meta manager
     StorageDir dir = mMetaManager.getTier("MEM").getDir(0);
     TempBlockMeta tempBlockMeta1 =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
     TempBlockMeta tempBlockMeta2 =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir);
     mMetaManager.addTempBlockMeta(tempBlockMeta1);
     mMetaManager.addTempBlockMeta(tempBlockMeta2);
 
@@ -267,9 +269,9 @@ public final class BlockMetadataManagerTest {
     StorageDir dir1 = mMetaManager.getTier("HDD").getDir(0);
     StorageDir dir2 = mMetaManager.getTier("HDD").getDir(1);
     TempBlockMeta tempBlockMeta1 =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir1);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir1);
     TempBlockMeta tempBlockMeta2 =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir2);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir2);
     mMetaManager.addTempBlockMeta(tempBlockMeta1);
     mMetaManager.addTempBlockMeta(tempBlockMeta2);
 
@@ -300,9 +302,9 @@ public final class BlockMetadataManagerTest {
     StorageDir dir2 = mMetaManager.getTier("HDD").getDir(1);
     long maxHddDir1Capacity = TIER_CAPACITY_BYTES[1][0];
     long blockMetaSize = maxHddDir1Capacity + 1;
-    BlockMeta blockMeta = new BlockMeta(TEST_BLOCK_ID, blockMetaSize, dir2);
+    BlockMeta blockMeta = new DefaultBlockMeta(TEST_BLOCK_ID, blockMetaSize, dir2);
     TempBlockMeta tempBlockMeta2 =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir1);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID2, TEST_BLOCK_SIZE, dir1);
     mMetaManager.addTempBlockMeta(tempBlockMeta2);
     dir2.addBlockMeta(blockMeta);
 
@@ -319,7 +321,7 @@ public final class BlockMetadataManagerTest {
   public void moveBlockMetaDeprecated() throws Exception {
     StorageDir dir = mMetaManager.getTier("MEM").getDir(0);
     TempBlockMeta tempBlockMeta =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
     mMetaManager.addTempBlockMeta(tempBlockMeta);
     mMetaManager.commitTempBlockMeta(tempBlockMeta);
     BlockMeta blockMeta = mMetaManager.getBlockMeta(TEST_TEMP_BLOCK_ID);
@@ -345,7 +347,7 @@ public final class BlockMetadataManagerTest {
   @Test
   public void moveBlockMetaDeprecatedExceedCapacity() throws Exception {
     StorageDir dir = mMetaManager.getTier("HDD").getDir(0);
-    BlockMeta blockMeta = new BlockMeta(TEST_BLOCK_ID, 2000, dir);
+    BlockMeta blockMeta = new DefaultBlockMeta(TEST_BLOCK_ID, 2000, dir);
     dir.addBlockMeta(blockMeta);
 
     mThrown.expect(WorkerOutOfSpaceException.class);
@@ -360,7 +362,7 @@ public final class BlockMetadataManagerTest {
   public void resizeTempBlockMeta() throws Exception {
     StorageDir dir = mMetaManager.getTier("MEM").getDir(0);
     TempBlockMeta tempBlockMeta =
-        new TempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_TEMP_BLOCK_ID, TEST_BLOCK_SIZE, dir);
     mMetaManager.resizeTempBlockMeta(tempBlockMeta, TEST_BLOCK_SIZE + 1);
     assertEquals(TEST_BLOCK_SIZE + 1, tempBlockMeta.getBlockSize());
   }
@@ -377,12 +379,12 @@ public final class BlockMetadataManagerTest {
     final long sessionId1 = 100;
     final long sessionId2 = 200;
     TempBlockMeta tempBlockMeta1 =
-        new TempBlockMeta(sessionId1, tempBlockId1, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(sessionId1, tempBlockId1, TEST_BLOCK_SIZE, dir);
     TempBlockMeta tempBlockMeta2 =
-        new TempBlockMeta(sessionId1, tempBlockId2, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(sessionId1, tempBlockId2, TEST_BLOCK_SIZE, dir);
     TempBlockMeta tempBlockMeta3 =
-        new TempBlockMeta(sessionId2, tempBlockId3, TEST_BLOCK_SIZE, dir);
-    BlockMeta blockMeta = new BlockMeta(TEST_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+        new DefaultTempBlockMeta(sessionId2, tempBlockId3, TEST_BLOCK_SIZE, dir);
+    BlockMeta blockMeta = new DefaultBlockMeta(TEST_BLOCK_ID, TEST_BLOCK_SIZE, dir);
     dir.addTempBlockMeta(tempBlockMeta1);
     dir.addTempBlockMeta(tempBlockMeta2);
     dir.addTempBlockMeta(tempBlockMeta3);
