+++ core/server/worker/src/test/java/alluxio/worker/block/BlockWorkerTest.java
@@ -36,6 +36,8 @@ import alluxio.underfs.UfsManager;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.util.io.PathUtils;
 import alluxio.worker.block.meta.BlockMeta;
+import alluxio.worker.block.meta.DefaultBlockMeta;
+import alluxio.worker.block.meta.DefaultTempBlockMeta;
 import alluxio.worker.block.meta.StorageDir;
 import alluxio.worker.block.meta.TempBlockMeta;
 import alluxio.worker.file.FileSystemMasterClient;
@@ -60,7 +62,7 @@ import java.util.Set;
  */
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({BlockMasterClient.class, BlockMasterClientPool.class, FileSystemMasterClient.class,
-    BlockHeartbeatReporter.class, BlockMetricsReporter.class, BlockMeta.class,
+    BlockHeartbeatReporter.class, BlockMetricsReporter.class, DefaultBlockMeta.class,
     BlockStoreLocation.class, StorageDir.class, ServerConfiguration.class, UnderFileSystem.class,
     BlockWorker.class, Sessions.class})
 public class BlockWorkerTest {
@@ -174,7 +176,7 @@ public class BlockWorkerTest {
     String mediumType = "MEM";
     HashMap<String, Long> usedBytesOnTiers = new HashMap<>();
     usedBytesOnTiers.put(tierAlias, usedBytes);
-    BlockMeta blockMeta = PowerMockito.mock(BlockMeta.class);
+    BlockMeta blockMeta = PowerMockito.mock(DefaultBlockMeta.class);
     BlockStoreLocation blockStoreLocation = PowerMockito.mock(BlockStoreLocation.class);
     BlockStoreMeta blockStoreMeta = mock(BlockStoreMeta.class);
 
@@ -208,7 +210,7 @@ public class BlockWorkerTest {
     String tierAlias = "MEM";
     HashMap<String, Long> usedBytesOnTiers = new HashMap<>();
     usedBytesOnTiers.put(tierAlias, usedBytes);
-    BlockMeta blockMeta = PowerMockito.mock(BlockMeta.class);
+    BlockMeta blockMeta = PowerMockito.mock(DefaultBlockMeta.class);
     BlockStoreLocation blockStoreLocation = PowerMockito.mock(BlockStoreLocation.class);
     BlockStoreMeta blockStoreMeta = mock(BlockStoreMeta.class);
 
@@ -236,7 +238,7 @@ public class BlockWorkerTest {
     String tierAlias = "MEM";
     BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
     StorageDir storageDir = mock(StorageDir.class);
-    TempBlockMeta meta = new TempBlockMeta(sessionId, blockId, initialBytes, storageDir);
+    TempBlockMeta meta = new DefaultTempBlockMeta(sessionId, blockId, initialBytes, storageDir);
 
     when(mBlockStore.createBlock(sessionId, blockId,
         AllocateOptions.forCreate(initialBytes, location))).thenReturn(meta);
@@ -259,7 +261,7 @@ public class BlockWorkerTest {
     String tierAlias = "HDD";
     BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
     StorageDir storageDir = mock(StorageDir.class);
-    TempBlockMeta meta = new TempBlockMeta(sessionId, blockId, initialBytes, storageDir);
+    TempBlockMeta meta = new DefaultTempBlockMeta(sessionId, blockId, initialBytes, storageDir);
 
     when(mBlockStore.createBlock(sessionId, blockId,
         AllocateOptions.forCreate(initialBytes, location))).thenReturn(meta);
@@ -281,7 +283,7 @@ public class BlockWorkerTest {
     String tierAlias = "MEM";
     BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
     StorageDir storageDir = mock(StorageDir.class);
-    TempBlockMeta meta = new TempBlockMeta(sessionId, blockId, initialBytes, storageDir);
+    TempBlockMeta meta = new DefaultTempBlockMeta(sessionId, blockId, initialBytes, storageDir);
 
     when(mBlockStore.createBlock(sessionId, blockId,
         AllocateOptions.forCreate(initialBytes, location))).thenReturn(meta);
@@ -392,7 +394,7 @@ public class BlockWorkerTest {
     BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
     BlockStoreLocation existingLocation = mock(BlockStoreLocation.class);
     when(existingLocation.belongsTo(location)).thenReturn(false);
-    BlockMeta meta = mock(BlockMeta.class);
+    BlockMeta meta = mock(DefaultBlockMeta.class);
     when(meta.getBlockLocation()).thenReturn(existingLocation);
     when(mBlockStore.getBlockMeta(eq(sessionId), eq(blockId), anyLong()))
         .thenReturn(meta);
@@ -413,7 +415,7 @@ public class BlockWorkerTest {
     BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
     BlockStoreLocation existingLocation = mock(BlockStoreLocation.class);
     when(existingLocation.belongsTo(location)).thenReturn(true);
-    BlockMeta meta = mock(BlockMeta.class);
+    BlockMeta meta = mock(DefaultBlockMeta.class);
     when(meta.getBlockLocation()).thenReturn(existingLocation);
     when(mBlockStore.getBlockMeta(eq(sessionId), eq(blockId), anyLong()))
         .thenReturn(meta);
@@ -433,7 +435,7 @@ public class BlockWorkerTest {
     long blockSize = mRandom.nextLong();
     StorageDir storageDir = mock(StorageDir.class);
     when(storageDir.getDirPath()).thenReturn("/tmp");
-    BlockMeta meta = new BlockMeta(blockId, blockSize, storageDir);
+    BlockMeta meta = new DefaultBlockMeta(blockId, blockSize, storageDir);
     when(mBlockStore.getBlockMeta(sessionId, blockId, lockId)).thenReturn(meta);
 
     mBlockWorker.readBlock(sessionId, blockId, lockId);
