+++ core/server/worker/src/test/java/alluxio/worker/block/meta/DefaultTempBlockMetaTest.java
@@ -11,7 +11,6 @@
 
 package alluxio.worker.block.meta;
 
-import alluxio.util.io.BufferUtils;
 import alluxio.util.io.PathUtils;
 import alluxio.worker.block.TieredBlockStoreTestUtils;
 
@@ -21,12 +20,10 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import java.io.IOException;
-
 /**
- * Unit tests for {@link BlockMeta}.
+ * Unit tests for {@link DefaultTempBlockMeta}.
  */
-public class BlockMetaTest {
+public class DefaultTempBlockMetaTest {
   private static final long TEST_SESSION_ID = 2;
   private static final long TEST_BLOCK_ID = 9;
   private static final long TEST_BLOCK_SIZE = 100;
@@ -34,11 +31,10 @@ public class BlockMetaTest {
   private static final String TEST_TIER_ALIAS = "MEM";
   private static final long[] TEST_TIER_CAPACITY_BYTES = {100};
   private static final String[] TEST_TIER_MEDIUM_TYPES = {"MEM"};
-  private static final String TEST_WORKER_DIR = "testworker";
-  private BlockMeta mBlockMeta;
-  private TempBlockMeta mTempBlockMeta;
+  private static final String TEST_WORKER_DATA_FOLDER = "workertest";
   private String mTestDirPath;
   private String mTestBlockDirPath;
+  private TempBlockMeta mTempBlockMeta;
 
   /** Rule to create a new temporary folder during each test. */
   @Rule
@@ -53,43 +49,52 @@ public class BlockMetaTest {
     // Sets up tier with one storage dir under mTestDirPath with 100 bytes capacity.
     TieredBlockStoreTestUtils.setupConfWithSingleTier(null, TEST_TIER_ORDINAL,
         TEST_TIER_ALIAS, new String[] {mTestDirPath}, TEST_TIER_CAPACITY_BYTES,
-        TEST_TIER_MEDIUM_TYPES, TEST_WORKER_DIR);
+        TEST_TIER_MEDIUM_TYPES, TEST_WORKER_DATA_FOLDER);
 
-    mTestBlockDirPath = PathUtils.concatPath(mTestDirPath, TEST_WORKER_DIR);
     StorageTier tier = StorageTier.newStorageTier(TEST_TIER_ALIAS, false);
     StorageDir dir = tier.getDir(0);
-    mTempBlockMeta = new TempBlockMeta(TEST_SESSION_ID, TEST_BLOCK_ID, TEST_BLOCK_SIZE, dir);
+    // Append the worker data folder to the expected path.
+    mTestBlockDirPath = PathUtils.concatPath(mTestDirPath, TEST_WORKER_DATA_FOLDER);
+    mTempBlockMeta = new DefaultTempBlockMeta(TEST_SESSION_ID, TEST_BLOCK_ID, TEST_BLOCK_SIZE, dir);
   }
 
   /**
-   * Tests the {@link BlockMeta#getBlockSize()} method.
-   */
-  @Test
-  public void getBlockSize() throws IOException {
-    // With the block file not really existing, expect committed block size to be zero.
-    mBlockMeta = new BlockMeta(mTempBlockMeta);
-    Assert.assertEquals(0, mBlockMeta.getBlockSize());
-
-    // With the block file partially written, expect committed block size equals real file size.
-    byte[] buf = BufferUtils.getIncreasingByteArray((int) TEST_BLOCK_SIZE - 1);
-    BufferUtils.writeBufferToFile(mTempBlockMeta.getCommitPath(), buf);
-    mBlockMeta = new BlockMeta(mTempBlockMeta);
-    Assert.assertEquals(TEST_BLOCK_SIZE - 1, mBlockMeta.getBlockSize());
-
-    // With the block file fully written, expect committed block size equals target block size.
-    buf = BufferUtils.getIncreasingByteArray((int) TEST_BLOCK_SIZE);
-    BufferUtils.writeBufferToFile(mTempBlockMeta.getCommitPath(), buf);
-    mBlockMeta = new BlockMeta(mTempBlockMeta);
-    Assert.assertEquals(TEST_BLOCK_SIZE, mBlockMeta.getBlockSize());
-  }
-
-  /**
-   * Tests the {@link BlockMeta#getPath()} method.
+   * Tests the {@link DefaultTempBlockMeta#getPath()} method.
    */
   @Test
   public void getPath() {
-    mBlockMeta = new BlockMeta(mTempBlockMeta);
+    Assert.assertEquals(PathUtils.concatPath(mTestBlockDirPath,
+        ".tmp_blocks", TEST_SESSION_ID % 1024,
+        String.format("%x-%x", TEST_SESSION_ID, TEST_BLOCK_ID)),
+        mTempBlockMeta.getPath());
+  }
+
+  /**
+   * Tests the {@link DefaultTempBlockMeta#getCommitPath()} method.
+   */
+  @Test
+  public void getCommitPath() {
     Assert.assertEquals(PathUtils.concatPath(mTestBlockDirPath, TEST_BLOCK_ID),
-        mBlockMeta.getPath());
+        mTempBlockMeta.getCommitPath());
+  }
+
+  /**
+   * Tests the {@link DefaultTempBlockMeta#getSessionId()} method.
+   */
+  @Test
+  public void getSessionId() {
+    Assert.assertEquals(TEST_SESSION_ID, mTempBlockMeta.getSessionId());
+  }
+
+  /**
+   * Tests the {@link DefaultTempBlockMeta#setBlockSize(long)} method.
+   */
+  @Test
+  public void setBlockSize() {
+    Assert.assertEquals(TEST_BLOCK_SIZE, mTempBlockMeta.getBlockSize());
+    mTempBlockMeta.setBlockSize(1);
+    Assert.assertEquals(1, mTempBlockMeta.getBlockSize());
+    mTempBlockMeta.setBlockSize(100);
+    Assert.assertEquals(100, mTempBlockMeta.getBlockSize());
   }
 }
